:recycle: Line, Sections에 구간 제거 메서드 추가
